python_functions/pyproject.toml
[project]
name = "python-functions"
version = "0.1.0"
description = "Python Lambda functions for ai-project"
authors = [
    {name = "Developer", email = "developer@example.com" },
]
requires-python = "==3.12.*"
dependencies = [
    "boto3",
    "requests",
    "sst",
    "strands-agents",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
sst = { git = "https://github.com/sst/sst.git", subdirectory = "sdk/python", rev = "dev" }
